+++ OKX.Api/Common/Enums/OkxInstrumentState.cs
@@ -18,8 +18,15 @@ public enum OkxInstrumentState : byte
     Suspend = 2,
 
     /// <summary>
+    /// Rebase
+    /// Can't be traded during rebasing. Only applicable to SWAP.
+    /// </summary>
+    [Map("rebase")]
+    Rebase = 3,
+
+    /// <summary>
     /// PreOpen
     /// </summary>
     [Map("preopen")]
-    PreOpen = 3,
+    PreOpen = 4,
 }